refactor: 테스트 픽스처 리팩토링
